fix: remove trailing "/" in `get_public_url`
When using the `get_public_url` methods, you get the following
responses:

```
https://SUPABASE_ID.supabase.co/storage/v1//object/public/BUCKET/FILE
```

Notice the double slash between
the `v1` and the `object` substrings in the path.

Even though the URL works---as
captured by the tests--it is a non
coherent way of rendering the
URLs.

This commit sets it so that all the
sub-paths inside the public URL
have a single slash.

Authored-By: Diego Rodriguez